pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-analyzer"
version = "2.0.0"
description = "Analyze your chess opening mistakes and train to fix them"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "python-chess>=1.999",
    "python-multipart>=0.0.6",
    "requests>=2.28",
]

[project.scripts]
chess-analyzer = "chess_analyzer.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["chess_analyzer*"]

[tool.setuptools.package-data]
chess_analyzer = ["static/*", "static/**/*"]
